Support for filtering array items (#60)
* Support for filtering array items

The JToken.SelectToken(string path) expects a unique path, but all array items have the same path, for example "policy[*].processId"

The JToken.SelectToken will throw an exception in case more than one token is found for a path. 

Therefore, the JToken.SelectTokens(string path) should be used. This returns an IEnumerable<JToken>?. If there are any tokens, all tokens for this path are removed. To stay in the example I gave above, when pathToRemove equals "policy[*].processId", the "processId" property for all policies will be removed now and no exception is thrown.

* Update FilterBySchemaTests.cs