implementation of path functions and indirect to use in the path and as function to easily reference the original locations (#101)
* Enhance JsonPath handling and improve test coverage

- Updated `JsonPathItemsFetcher` to handle null data gracefully.
- Fixed typo in `SelectedTokensExtensions` class name.
- Simplified `JLio.UnitTests.csproj` by using wildcards for test data files.
- Added `JsonPathMethodsEdgeCasesTests` for edge case validation.
- Expanded `JsonPathMethodsTests` with additional scenarios and performance checks.
- Created/updated test data files for comprehensive testing.

* indirect in the paths are working

* Add ScriptPath function and class for path handling

Updated the ParseOptions class to register a new function, ScriptPath, alongside the existing Indirect function.

Introduced a new ScriptPath class in ScriptPath.cs that inherits from FunctionBase. This class provides constructors for path initialization and an Execute method to process tokens and return paths. It supports current item references and absolute paths, enhancing path handling capabilities. A private method, GetTokenPath, is also included to retrieve the path of a given token.

* Add parent navigation support to IItemsFetcher

- Introduced `ParentPathIndicator` to `IItemsFetcher` and `JsonPathItemsFetcher` for parent path navigation in JSON.
- Enhanced `FixedValue` class to manage parent item navigation using the new indicator.
- Implemented unit tests in `JTokenTests.cs` to validate parent navigation functionality across various scenarios and edge cases.

* Enhance FunctionBase and ScriptPath functionality

- Made FunctionName property virtual in FunctionBase for overriding.
- Updated ScriptPath to include necessary namespaces and override FunctionName to return "path".
- Modified targetPath validation to check for empty strings and `''`.
- Added ScriptPathTests with multiple unit tests for various scenarios.
- Updated and added JSON files for test data and expected results.

* Add support for parent path indicator in ScriptPath

Updated the `ScriptPath` class to handle the new parent path indicator `@.<--`, allowing retrieval of the current item's path. Added a corresponding test method in `ScriptPathTests` to validate this functionality. Created new JSON files for test data, expected output, and a script demonstrating the use of the new syntax.